+++ arcore_plugin/example/lib/main.dart
@@ -50,13 +50,18 @@ class _TextViewExampleState extends State<TextViewExample> {
           centerTitle: true,
         ),
         body: Center(
-            child: Container(
-                width: screenSize.width,
-                height: screenSize.height,
-                child: ArCoreView(
-                  onImageRecognized: _onImageRecognized,
-                  onArCoreViewCreated: _onTextViewCreated,
-                ))));
+            child: ArCoreView(
+          focusBox: Container(
+            width: screenSize.width * 0.5,
+            height: screenSize.width * 0.5,
+            decoration: BoxDecoration(
+                border: Border.all(width: 1, style: BorderStyle.solid)),
+          ),
+          width: screenSize.width,
+          height: screenSize.height,
+          onImageRecognized: _onImageRecognized,
+          onArCoreViewCreated: _onTextViewCreated,
+        )));
   }
 
   void _onTextViewCreated(ArCoreViewController controller) {